Added the repack line which actually makes it work

+++ Move Files in this to a preferred space/edit_launcher_carousel.py
@@ -18,8 +18,8 @@ shutil.copyfile(launcher_path / "app.asar", launcher_path / "original_app.asar")
 
 
 #The command to unpack the launcher file
-npx_command = ["npx", "asar", "extract", "app.asar", "unpacked"]
-# npx_command = ["echo", "Hello World!"]  # for testing subprocess
+npx_command_extract = ["npx", "asar", "extract", "app.asar", "unpacked"]
+# npx_command_extract = ["echo", "Hello World!"]  # for testing subprocess
 
 # Replace this with the path to your RSI Launcher
 path = Path(r"C:\Program Files\Roberts Space Industries\RSI Launcher\resources")
@@ -31,7 +31,7 @@ if not path.exists():
 # check if the unpacked folder exists and skip to the next step if it does
 if not (path / "unpacked").exists():
     # Run the command
-    process = subprocess.Popen(npx_command, cwd=path, shell=True)
+    process = subprocess.Popen(npx_command_extract, cwd=path, shell=True)
 
     # Check the return code of the command
     if process.wait() == 0:
@@ -52,7 +52,8 @@ with open(new_path, "r") as f:
     data = f.read()
 
 #rename the original javascript file, so we have it just incase
-os.rename(new_path, path / "unpacked" / "app"/"original_cig-launcher.js")
+if not (path / "unpacked" / "app"/"original_cig-launcher.js").exists():
+    os.rename(new_path, path / "unpacked" / "app"/"original_cig-launcher.js")
 
 # we are looking for this line carousel:{delay:25e3,images:[ this is where the list of image names starts
 pattern = re.compile(r'carousel:{delay:25e3,images:\[([^]]*)\]')
@@ -72,12 +73,32 @@ new_data = data.replace(current_img_list, replacement_string)
 
 
 # save the original images
-shutil.copytree(launcher_images, path / "unpacked" / "app"/"assets"/"original_images")
+if not (path / "unpacked" / "app"/"assets"/"original_images").exists():
+    shutil.copytree(launcher_images, path / "unpacked" / "app"/"assets"/"original_images")
 #remove the original images
 shutil.rmtree(launcher_images)
 #finally lets copy the new images to the correct folder
 shutil.copytree(image_folder, launcher_images)
 
+with open(new_path, "w") as f:
+    f.write(new_data)
+
 #if you want to delete the unpacked folder after you are done remove the "#" from the line below
 # shutil.rmtree(path / "unpacked")
 
+#The command to unpack the launcher file
+npx_command_pack = ["npx", "asar", "pack","unpacked", "app.asar"]
+# npx_command_extract = ["echo", "Hello World!"]  # for testing subprocess
+
+# Replace this with the path to your RSI Launcher
+
+process = subprocess.Popen(npx_command_pack, cwd=path, shell=True)
+
+# Check the return code of the command
+if process.wait() == 0:
+    print('npm command succeeded.')
+else:
+    print('npm command failed. New images have not been applied to the launcher.')
+
+    # Print the output of the command for debugging
+    print(process)
